backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "python-multipart>=0.0.24",
    "scikit-image>=0.26.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.44.0",
]
